+++ src/generators/saw_generator.h
@@ -7,7 +7,7 @@
 class SawGenerator: public Generator
 {
     public:
-    float getSample();
+    int getSample();
     void  initLUT(int size);
 };
