+++ src/generators/sine_generator.h
@@ -7,7 +7,7 @@
 class SineGenerator: public Generator
 {
     public:
-    float getSample();
+    int getSample();
     void  initLUT(int size);
 };
