+++ src/generators/square_generator.h
@@ -7,7 +7,7 @@
 class SquareGenerator: public Generator
 {
     public:
-    float getSample();
+    int getSample();
     void  initLUT(int size);
 };
